feat: replaced header buttons with icons
