+++ package.json
@@ -13,9 +13,9 @@
     "coverage": "jest --coverage"
   },
   "dependencies": {
-    "@radix-ui/react-separator": "^1.0.3",
     "@hookform/resolvers": "^3.3.2",
     "@radix-ui/react-label": "^2.0.2",
+    "@radix-ui/react-separator": "^1.0.3",
     "@radix-ui/react-slot": "^1.0.2",
     "@radix-ui/react-switch": "^1.0.3",
     "@radix-ui/react-toast": "^1.1.5",
