+++ src/components/Header/Header.tsx
@@ -1,12 +1,13 @@
 import { useEffect, useState } from 'react';
 import Link from 'next/link';
-import { Button, buttonVariants } from '../ui/button';
+import { Button } from '../ui/button';
 import { Switch } from '../ui/switch';
 import { useAuthState } from 'react-firebase-hooks/auth';
 import { auth } from '@/firebase';
 import { signout } from '@/firebase/signOut';
 import { useLang } from '@/lib/langContext';
 import { textContent, TextContentType } from '@/lib/langText';
+import Image from 'next/image';
 
 const Header = () => {
   const [isScrolled, setIsScrolled] = useState(false);
@@ -31,13 +32,12 @@ const Header = () => {
       }`}
       data-testid="header"
     >
-      <div>
-        <Link href="/" className="mr-5">
-          {text.home}
-        </Link>
-      </div>
-      <div className="flex items-center gap-8">
-        <span className="flex items-center gap-2">
+      <Link href="/">
+        <Image src="/home.svg" width={100} height={100} alt="Logo" />
+      </Link>
+
+      <div className="flex items-center gap-8 pl-4">
+        <span className="flex flex-col items-center gap-2 min-[400px]:flex-row">
           {text.en}
           <Switch
             name="switcher"
@@ -49,26 +49,26 @@ const Header = () => {
         </span>
         {user ? (
           <Button
-            variant="secondary"
+            className="bg-slate-800 hover:bg-slate-600 rounded h-10 w-10 p-1 duration-150"
             onClick={() => {
               signout();
             }}
           >
-            {text.signOut}
+            <Image src="/exit.svg" width={25} height={25} alt="Exit" />
           </Button>
         ) : (
           <span className="flex flex-wrap justify-end gap-1">
             <Link
               href={'/login'}
-              className={buttonVariants({ variant: 'secondary', size: 'sm' })}
+              className="bg-slate-800 hover:bg-slate-600 rounded h-10 w-10 p-1 duration-150"
             >
-              {text.login}
+              <Image src="/signin.svg" width={30} height={30} alt="Sign in" />
             </Link>
             <Link
               href={'/register'}
-              className={buttonVariants({ variant: 'secondary', size: 'sm' })}
+              className="bg-slate-800 hover:bg-slate-600 rounded h-10 w-10 p-1 duration-150"
             >
-              {text.register}
+              <Image src="/signup.svg" width={28} height={28} alt="Register" />
             </Link>
           </span>
         )}
