+++ src/test/header.test.tsx
@@ -3,7 +3,6 @@ import '@testing-library/jest-dom';
 import ReactDOM from 'react-dom/client';
 import { act } from 'react-dom/test-utils';
 import Header from '@/components/Header/Header';
-import { LangProvider } from '@/lib/langContext';
 
 let container: HTMLDivElement | null;
 let header: HTMLElement;
@@ -14,11 +13,7 @@ describe('Tests for header', () => {
     document.body.appendChild(container);
     await act(async () => {
       if (container) {
-        ReactDOM.createRoot(container).render(
-          <LangProvider>
-            <Header />
-          </LangProvider>
-        );
+        ReactDOM.createRoot(container).render(<Header />);
       }
     });
     header = screen.getByTestId('header');
@@ -52,17 +47,5 @@ describe('Tests for header', () => {
       fireEvent.scroll(window, { target: { scrollY: 0 } });
     });
     expect(header).not.toHaveClass('pt-2', 'pb-2');
-  });
-
-  test('Header changes language after clicking the switch', () => {
-    const switcher = screen.getByTestId('switch-lang');
-    const buttons = screen.getAllByRole('link');
-    expect(switcher).not.toBeChecked();
-    expect(buttons[1]).toHaveTextContent('Login');
-    act(() => {
-      fireEvent.click(switcher);
-    });
-    expect(switcher).toBeChecked();
-    expect(buttons[1]).toHaveTextContent('Войти');
   });
 });
