+++ src/test/welcome.test.tsx
@@ -1,8 +1,10 @@
-import { screen, waitFor } from '@testing-library/react';
+import { screen, waitFor, fireEvent } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import ReactDOM from 'react-dom/client';
 import { act } from 'react-dom/test-utils';
 import Welcome from '@/components/Welcome';
+import Header from '@/components/Header/Header';
+import { LangProvider } from '@/lib/langContext';
 
 let container: HTMLDivElement | null;
 
@@ -12,7 +14,14 @@ describe('Welcome page', () => {
     document.body.appendChild(container);
     await act(async () => {
       if (container) {
-        ReactDOM.createRoot(container).render(<Welcome />);
+        ReactDOM.createRoot(container).render(
+          <>
+            <LangProvider>
+              <Header />
+              <Welcome />
+            </LangProvider>
+          </>
+        );
       }
     });
   });
@@ -45,5 +54,17 @@ describe('Welcome page', () => {
 
   test('Main route exists', () => {
     expect(global.location.pathname).toBe('/');
+  });
+
+  test('Welcome page changes language after clicking the switch', () => {
+    const switcher = screen.getByTestId('switch-lang');
+    const links = screen.getAllByRole('link');
+    expect(switcher).not.toBeChecked();
+    expect(links[3]).toHaveTextContent('Login');
+    act(() => {
+      fireEvent.click(switcher);
+    });
+    expect(switcher).toBeChecked();
+    expect(links[3]).toHaveTextContent('Войти');
   });
 });
